Refactor: ADD UserService interface 추가
- 인터페이스 추가 후 테스트코드 작성.
